requirements.txt
Django==2.0.2
djangorestframework==3.7.7
numpy==1.11.0
python-dateutil==2.6.1
pytz==2018.3
scipy==0.17.0
six==1.11.0
treetaggerwrapper==2.2.3
typing==3.6.4
